Add support for arm64 Object files with target > iOS 12
