+++ Sources/Tests/TestResources/return2.c.fixture
@@ -1,0 +1,3 @@
+int return2() {
+  return 2;
+}
